refactor: split RuleTranslator's 782-line monolith into per-adapter modules
Closes #196. Every provider's rule-translation logic (both directions,
both providers) lived in one class in one file. Every new provider would
add its own pair of directions to the same shared file — the same shape
of problem the credential duplication was (#182/#193), just for rule
translation instead of credentials, and heading toward 1500+ lines across
Fastly/GCP/AWS if left alone.

Split into src/lib/providers/vercel/translator.ts and .../cloudflare/
translator.ts, each holding only that provider's Unified<->native logic
and living next to that provider's client/service/credentials. RuleTranslator
itself becomes a ~35-line static-method facade delegating to the two new
modules, so none of RuleTranslator's ~15 existing call sites (Cloudflare/
VercelFirewallService, vercelConfigAdapter, ExpressionBuilder's comments)
needed to change.

Shared types (TranslationWarning, TranslationResult, etc.) move to a new
TranslationTypes.ts, which also resolves a small pre-existing wart:
TranslationWarningSystem.ts imported these types from RuleTranslator.ts
while RuleTranslator.ts's own methods lazily require()'d
TranslationWarningSystem at call time specifically to dodge the circular
import that direction would otherwise create. Both files now depend on
TranslationTypes.ts instead of each other, so the lazy requires became
ordinary top-level imports.

Also removes RuleTranslator.vercelToCloudflare/cloudflareToVercel — direct
Vercel<->Cloudflare translation bypassing the Unified model. Repo-wide
search confirmed zero production callers (only their own tests exercised
them); doorman has never routed data provider-to-provider directly since
the Unified hub model landed, and keeping them would have forced either an
artificial cross-adapter import or a third catch-all file, undermining the
point of this split. Their window-parsing test coverage (parseWindowToSeconds,
still live via unifiedToCloudflare) was retargeted rather than lost.

While comparing the two Cloudflare-outputting functions before deleting the
dead one, found unifiedToCloudflare (the live path) is missing the
redirect-parameter-building logic the dead vercelToCloudflare had — filed
separately as #199 rather than fixed here, since this PR is a structural
move and deliberately preserves existing (including buggy) behavior.

Net: no single file holds more than one provider's translation logic;
adding a provider's translator means adding its own directory, not editing
an existing provider's file.
